Migrate to TypeScript
- Rename entry to index
- Move code and assets into src
- Add typescript definitions for Main.elm
