+++ package.json
@@ -2,15 +2,15 @@
   "name": "sparksp-timers",
   "version": "0.0.1",
   "description": "A collection of web timers for counting up and down.",
-  "main": "index.js",
+  "main": "src/index.ts",
   "scripts": {
     "clean:cache": "rimraf .cache",
     "clean:dist": "rimraf dist",
     "clean:elm": "rimraf elm-stuff",
     "clean": "run-p clean:cache clean:dist clean:elm",
     "prebuild": "run-s clean",
-    "build": "parcel build entry.html --out-file index.html",
-    "dev": "parcel entry.html --out-file index.html --open",
+    "build": "parcel build src/index.html --out-file index.html",
+    "dev": "parcel src/index.html --out-file index.html --open",
     "lint": "elm-review",
     "test": "elm-test"
   },
@@ -31,6 +31,7 @@
     "elm-hot": "^1.1.4",
     "elm-review": "^2.0.0",
     "elm-test": "^0.19.1-revision2",
+    "elm-types": "^0.0.2",
     "npm-run-all": "^4.1.5",
     "parcel-bundler": "^1.12.4",
     "parcel-plugin-purgecss": "^2.1.3",
@@ -39,7 +40,8 @@
     "postcss-modules": "^1.5.0",
     "regenerator-runtime": "^0.13.5",
     "rimraf": "^3.0.2",
-    "tailwindcss": "^1.2.0"
+    "tailwindcss": "^1.2.0",
+    "typescript": "^3.8.3"
   },
   "dependencies": {}
 }
