+++ src/Main.elm.d.ts
@@ -1,0 +1,4 @@
+import {Bundle} from 'elm-types'
+import {Main} from './Main'
+
+export const Elm: Bundle<{Main: Main}>
