+++ src/assets/img/android-icon-144x144.png
