+++ src/assets/img/apple-icon-precomposed.png
