+++ src/assets/img/ipad_splash.png
