+++ src/assets/img/ipadpro3_splash.png
