+++ src/assets/img/iphone5_splash.png
