+++ src/assets/img/iphone6_splash.png
