+++ src/assets/img/iphoneplus_splash.png
