+++ src/assets/img/iphonexr_splash.png
