+++ src/index.html
@@ -36,13 +36,6 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Timers</title>
     <link rel="manifest" href="manifest.webmanifest">
-    <script defer src="entry.js"></script>
 </head>
-<body ontouchstart="">
-<main>
-    <div class="flex items-center justify-center p-32">
-        <img src="assets/img/app-icon.png" alt="A clock counting down" class="max-h-full">
-    </div>
-</main>
-</body>
+<body ontouchstart=""><script defer src="index.ts"></script></body>
 </html>
