+++ static/apple-touch-icon.png
@@ -1,1 +1,1 @@
-../assets/img/app-icon.png
+../src/assets/img/app-icon.png
