+++ static/favicon.ico
@@ -1,1 +1,1 @@
-../assets/img/favicon.ico
+../src/assets/img/favicon.ico
